Move the not keyring related sync tests into `test_lockfile.py`
